requirements.txt
watchdog
psycopg2-binary
python-dotenv
requests
pydantic
